requirements.txt
asgiref==3.3.1
Django==3.1.4
django-phonenumber-field==5.0.0
mysqlclient==2.0.1
phonenumbers==8.12.14
pytz==2020.4
sqlparse==0.4.1
